refactor(pcc-dashboard-general-information): move files
