+++ client/app/dedicatedCloud/dashboard/tiles/general-information/general-information.controller.js
