+++ client/app/dedicatedCloud/dashboard/tiles/general-information/general-information.html
