+++ client/app/dedicatedCloud/dashboard/tiles/general-information/translations/Messages_fr_FR.json
